Fixup x86 conversion routines
Some time ago, s390 worked around similar issues by doing radix
conversions via string conversion. Simple, slow, and usually
more correct.

Shuffle the Implies around such that these can be used for other
architectures. Similarly, move the symbol information to avoid
exposing TFmode or XFmode variants where not implemented. Similarly,
split the Version files to avoid arch specific long double details
from polluting common Version files.

Likewise, s390 and ppc both use TFmode to represent long double
operands. x86 uses XFmode, and TFmode for _Float128. Today libdfp
makes no attempt to support directly support _Float128, thus x86
users get broken long double conversions. Therefore, we must update
the macros to use XFmode in place of TFmode.

Likewise, fixup FP_LLOGB* macros. They are wrong for x86.

Finally, cleanup error message printed for long int checks so the
output formats match. Notably, test-cast-underflow tests still fail.
There are accuracy issues at the extreme ends of conversion which
have been uncovered the by the tests added to verify s390 hardware
conversions.

+++ base-math/truncdddf.c
@@ -74,7 +74,7 @@ CONVERT_WRAPPER(
 	    switch (fegetround())
 	      {
 	        case FE_TONEAREST:
-	          mant = llabs (a_norm);
+	          mant = llabs ((long long)a_norm);
 	          if (exp < -POWOF10_MIN_DENORM_DBL_EXP
 	              || ((exp == -POWOF10_MIN_DENORM_DBL_EXP)
 	              && (mant >= (POWOF10_MIN_DENORM_DBL_MAN/2LL))))
